+++ src/runtime/types/index.d.ts
@@ -13,3 +13,21 @@ declare module '#app' {
 }
 
 type S3ObjectMetadata = Record<string, string>
+
+interface ModuleOptionsFS {
+  driver: 'fs';
+  fsBase?: string;
+  accept?: string;
+  maxSizeMb?: number;
+}
+
+interface ModuleOptionsS3 {
+  driver: 's3';
+  accessKeyId: string;
+  secretAccessKey: string;
+  endpoint: string;
+  region: string;
+  bucket: string;
+  accept?: string;
+  maxSizeMb?: number;
+}
